requirements.txt
flask
flask-cors
numpy
pandas
scikit-image
opencv-python
pillow
scipy
scikit-learn
nibabel
pydicom
joblib
